requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
SQLAlchemy==2.0.4
Werkzeug==2.2.3
Jinja2==3.1.2
click==8.1.6
itsdangerous==2.1.2
pandas==1.5.3
openpyxl==3.1.1
xlrd==2.0.1
Pillow==9.4.0
gunicorn==20.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.6  # PostgreSQL adapter
